pyproject.toml
[project]
name = "complexity"
version = "0.1.0"
requires-python = ">=3.10" 
dependencies = [
    "datasets",
    "pandas",
    "matplotlib",
    "seaborn",
    "python-dotenv",
    "litellm",
    "requests",
    "markdownify",
    "tenacity>=9.0.0",
    "loguru>=0.7.3",
    "spacy>=3.8.4",
    "python-arango>=8.1.4",
    "deepmerge>=2.0",
    "torch",
    "sentence-transformers>=4.1.0",
    "transformers @ git+https://github.com/huggingface/transformers.git",
    "nltk>=3.9.1",
    "pyperclip>=1.9.0",
    "json-repair>=0.35.0",
    "rapidfuzz>=3.11.0",
    "tabulate>=0.9.0",
    "einops>=0.8.1",
    "numpy>=2.2.2",
    "scikit-learn>=1.6.1",
    # "faiss-gpu>=1.7.2",
    # "faiss-cpu>=1.11.0",
    # "faiss-gpu-cu12[fix-cuda]",
    "qdrant-client>=1.14.2",
    "faiss-cpu>=1.11.0",
    "colorama>=0.4.6",
    "rich>=13.9.4",
    "gitingest>=0.1.4",
    "tiktoken>=0.8.0",
    "google-auth>=2.39.0",
    "google-auth-oauthlib>=1.2.2",
    "redis>=6.0.0",
    "markitdown>=0.1.1",
    "pytest-asyncio>=0.26.0",
]

[project.scripts]
complexity-cli = "complexity.cli:app"  # Point directly to Typer app object, not wrapper function


[project.optional-dependencies]
dev = ["ruff"]
test = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
markers = [
    "dependency: mark test to run after specified dependencies",
]

